cleanup and type fixes
