+++ packages/cli/package.json
@@ -23,10 +23,6 @@
     "@oclif/plugin-not-found": "^2.3.5",
     "@oclif/plugin-plugins": "^2.1.4",
     "@oclif/plugin-version": "^1.1.3",
-    "@polkadot/api": "^9.0.1",
-    "@polkadot/api-contract": "^9.0.1",
-    "@polkadot/keyring": "^10.1.2",
-    "@polkadot/util": "^10.1.2",
     "@polkadot/util-crypto": "^10.1.2",
     "bn.js": "^5.2.1",
     "chalk": "4",
@@ -51,8 +47,6 @@
   },
   "devDependencies": {
     "@oclif/test": "^2.2.8",
-    "@polkadot/api-augment": "^9.0.1",
-    "@polkadot/types": "^9.0.1",
     "@types/chai": "^4",
     "@types/debug": "^4.1.7",
     "@types/download": "^8.0.1",
@@ -125,8 +119,5 @@
     "shiden",
     "wasm"
   ],
-  "types": "dist/index.d.ts",
-  "workspaces": [
-    "packages/*"
-  ]
+  "types": "dist/index.d.ts"
 }
