+++ packages/cli/src/commands/account/create.ts
@@ -1,11 +1,14 @@
 import { Command, Flags } from "@oclif/core";
 import chalk = require("chalk");
 import { writeJSON } from "fs-extra";
-import { ChainAccount } from "@astar-network/swanky-core";
-import { ensureSwankyProject, getSwankyConfig } from "@astar-network/swanky-core";
-import { encrypt } from "@astar-network/swanky-core/dist/lib";
+import {
+  ensureSwankyProject,
+  getSwankyConfig,
+  ChainAccount,
+  encrypt,
+  AccountData,
+} from "@astar-network/swanky-core";
 import inquirer from "inquirer";
-import { AccountData } from "@astar-network/swanky-core";
 export class CreateAccount extends Command {
   static description = "Create a new dev account in config";
