+++ packages/cli/src/index.ts
@@ -1,2 +1,1 @@
-import "@polkadot/api-augment";
 export { run } from "@oclif/core";
