+++ packages/core/src/lib/index.ts
@@ -2,6 +2,7 @@ export * from "./account";
 export * from "./command-utils";
 export * as consts from "./consts";
 export * from "./crypto";
+export * from "./deploy-api";
 export * from "./nodeInfo";
 export * from "./spinner";
 export * from "./substrate-api";
